+++ apps/ledger-live-mobile/src/newArch/features/Accounts/components/AddAccountButton/index.tsx
@@ -1,8 +1,5 @@
-import { Icons, Text } from "@ledgerhq/native-ui";
 import React, { FC, useCallback, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Pressable } from "react-native";
-import styled from "styled-components/native";
 import AddAccountDrawer from "LLM/features/Accounts/screens/AddAccount";
 import { track } from "~/analytics";
 import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
@@ -12,20 +9,7 @@ import {
   useModularDrawerVisibility,
 } from "LLM/features/ModularDrawer";
 import { findCryptoCurrencyById } from "@ledgerhq/live-common/currencies/index";
-
-const StyledPressable = styled(Pressable)`
-  border-width: 1px;
-  border-style: dotted;
-  border-color: ${({ theme }) => theme.colors.opacityDefault.c10};
-  padding: 16px;
-  margin-vertical: 8px;
-  border-radius: 12px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  column-gap: 12px;
-`;
+import { AddAccountButton as AddAccountButtonComponent } from "@ledgerhq/native-ui/pre-ldls/components/index";
 
 type Props = {
   sourceScreenName: string;
@@ -73,19 +57,11 @@ const AddAccountButton: FC<Props> = ({ sourceScreenName, disabled, currency, onC
 
   return (
     <>
-      <StyledPressable
+      <AddAccountButtonComponent
+        label={t("addAccounts.addNewOrExisting")}
+        onClick={handleOpenAddAccountModal}
         disabled={disabled}
-        style={({ pressed }: { pressed: boolean }) => [
-          { opacity: pressed ? 0.5 : 1.0, marginVertical: 12 },
-          disabled && { opacity: 0.5 },
-        ]}
-        hitSlop={6}
-        onPress={handleOpenAddAccountModal}
-        testID="add-new-account-button"
-      >
-        <Text variant="large">{t("addAccounts.addNewOrExisting")}</Text>
-        <Icons.Plus size="S" color="neutral.c100" />
-      </StyledPressable>
+      />
       <AddAccountDrawer isOpened={isAddAccountModalOpen} onClose={handleCloseAddAccountModal} />
     </>
   );
