+++ apps/ledger-live-mobile/src/newArch/features/ModularDrawer/screens/AccountSelection/index.tsx
@@ -1,15 +1,16 @@
-import React, { useCallback, useEffect, useRef } from "react";
+import React, { useCallback, useRef } from "react";
 import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { AccountLike } from "@ledgerhq/types-live";
 import { FlatList } from "react-native";
-import { BottomSheetVirtualizedList, useBottomSheetInternal } from "@gorhom/bottom-sheet";
+import { BottomSheetVirtualizedList } from "@gorhom/bottom-sheet";
 import { AccountItem } from "@ledgerhq/native-ui/pre-ldls/index";
 import { TrackDrawerScreen, EVENTS_NAME } from "../../analytics";
-import AddAccountButton from "~/newArch/features/Accounts/components/AddAccountButton";
-import { useDetailedAccounts } from "../../hooks/useDetailAccount";
+import { useDetailedAccounts } from "../../hooks/useDetailedAccounts";
 import { WalletAPIAccount } from "@ledgerhq/live-common/wallet-api/types";
 import { Observable } from "rxjs";
 import { AccountUI } from "@ledgerhq/native-ui/pre-ldls/components/";
+import { AddAccountButton } from "@ledgerhq/native-ui/pre-ldls/components/index";
+import { useTranslation } from "react-i18next";
 
 export type AccountSelectionStepProps = {
   accounts$?: Observable<WalletAPIAccount[]>;
@@ -18,6 +19,7 @@ export type AccountSelectionStepProps = {
   flow: string;
   source: string;
   onAddNewAccount: () => void;
+  onClose: (() => void) | undefined;
 };
 
 const HEADER_HEIGHT = 64;
@@ -30,30 +32,28 @@ const AccountSelection = ({
   source,
   accounts$,
   onAddNewAccount,
+  onClose,
 }: Readonly<AccountSelectionStepProps>) => {
-  const { detailedAccounts, onAccountClick } = useDetailedAccounts(asset!, flow, source, accounts$);
-  const { shouldHandleKeyboardEvents } = useBottomSheetInternal();
+  const { detailedAccounts, onAccountSelected } = useDetailedAccounts(
+    asset!,
+    flow,
+    source,
+    onClose,
+    accounts$,
+  );
   const listRef = useRef<FlatList>(null);
+  const { t } = useTranslation();
 
   const renderItem = useCallback(
     ({ item }: { item: AccountUI }) => {
-      return <AccountItem account={item} onClick={() => onAccountClick(item)} />;
+      return <AccountItem account={item} onClick={() => onAccountSelected(item)} />;
     },
-    [onAccountClick],
+    [onAccountSelected],
   );
 
   const renderFooter = useCallback(() => {
-    return (
-      <AddAccountButton
-        onClick={() => onAddNewAccount}
-        sourceScreenName={EVENTS_NAME.MODULAR_ACCOUNT_SELECTION}
-      />
-    );
-  }, [onAddNewAccount]);
-
-  useEffect(() => {
-    shouldHandleKeyboardEvents.value = false;
-  }, [shouldHandleKeyboardEvents]);
+    return <AddAccountButton label={t("addAccounts.addNewOrExisting")} onClick={onAddNewAccount} />;
+  }, [onAddNewAccount, t]);
 
   return (
     <>
